Python code for three experiments on entropy loss
Python code for three experiments on entropy loss:
a.  The "fan-out":  How many different random passwords would we expect to be able to map down to the same user-friendly password as this one?
b.  The entropy loss in bits.
c.  How many randomly-generated lowercase characters would need to be added to get back at least the same amount of entropy that was lost.